Integrate Search and Presearch and fix bugs where CudaPointers were being freed prematurely
This change is big and combines a lot of modifications. A lot of CudaPointers now use unique_pointers in order not to go out of scope prematurely. The Presearcher and Searcher classes have been introduced. Lots of code has been moved around so that simulataneous cuda and cpp files can be compiled together with nvcc and gcc respecively. The main file also right now acts as a test for the searcher and presearcher before a proper unit test is added