Rehaul to structure and documentation
So this is a bit of a god commit which was done to update the entire
repository to use more modern tools such as clang-tidy and also have
easier documentation (mostly automated, especially for files).

We also updated thte output format to include the format name and
version number of that format.

A bug fix was made so that memory on the gpu is only allocated once and
kept throughout the lifetime of the process